icinga-sunsetting/requirements.txt
pandas
jq
python-box
pypuppetdb
click
